+++ src/Microsoft.PowerShell.ConsoleGuiTools/OutConsoleGridViewCmdletCommand.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Management.Automation;
 using System.Management.Automation.Internal;
 using Microsoft.PowerShell.OutGridView.Models;
@@ -12,8 +11,8 @@ using Microsoft.PowerShell.OutGridView.Models;
 namespace Microsoft.PowerShell.ConsoleGuiTools;
 
 /// <summary>
-///     Sends output to an interactive table in a separate console window. This class is invoked by PowerShell when the
-///     Out-ConsoleGridView cmdlet is called.
+///     Sends output to an interactive table using Terminal.Gui's TableView.
+///     Renders with native column headers, sizing, horizontal scrolling, and streaming support.
 /// </summary>
 [Cmdlet(VerbsData.Out, "ConsoleGridView")]
 [Alias("ocgv")]
@@ -24,8 +23,8 @@ public class OutConsoleGridViewCmdletCommand : PSCmdlet, IDisposable
     private const string DATA_NOT_QUALIFIED_FOR_GRID_VIEW = nameof(DATA_NOT_QUALIFIED_FOR_GRID_VIEW);
     private const string ENVIRONMENT_NOT_SUPPORTED_FOR_GRID_VIEW = nameof(ENVIRONMENT_NOT_SUPPORTED_FOR_GRID_VIEW);
 
-    private readonly List<PSObject> _psObjects = [];
     private readonly OutConsoleGridView _outConsoleGridView = new();
+    private bool _initialized;
 
     #endregion Properties
 
@@ -41,7 +40,7 @@ public class OutConsoleGridViewCmdletCommand : PSCmdlet, IDisposable
     ///     Gets or sets the title of the Out-ConsoleGridView window.
     /// </summary>
     [Parameter(HelpMessage =
-        "Specifies the text that appears title bar of the Out-ConsoleGridView window. By default, the title displays Out-ConsoleGridView.")]
+        "Specifies the text that appears in the title bar of the Out-ConsoleGridView window. By default, the title bar displays the command that invokes Out-ConsoleGridView.")]
     [ValidateNotNullOrEmpty]
     public string? Title { get; set; }
 
@@ -63,23 +62,34 @@ public class OutConsoleGridViewCmdletCommand : PSCmdlet, IDisposable
     /// <summary>
     ///     Gets or sets a value indicating whether "minimum UI" mode will be enabled.
     /// </summary>
-    [Parameter(HelpMessage = "If specified no title or status bar will be displayed in the TUI. The filter will only be displayed if -Filter is specified.")]
+    [Parameter(HelpMessage = "If specified no window frame, filter box, or status bar will be displayed in the TUI.")]
     public SwitchParameter MinUI { set; get; }
+
+    /// <summary>
+    ///     Gets or sets a value indicating whether the TUI should run in full screen mode.
+    /// </summary>
+    [Parameter(HelpMessage = "If specified the TUI will run in full screen mode instead of inline.")]
+    public SwitchParameter FullScreen { set; get; }
+
+    /// <summary>
+    ///     Gets or sets which UI element should receive initial focus (Table or Filter).
+    /// </summary>
+    [Parameter(HelpMessage = "Specifies which UI element receives initial focus. Valid values are 'Table' (default) and 'Filter'.")]
+    public FocusTarget Focus { set; get; } = FocusTarget.Table;
+
+    /// <summary>
+    ///     Gets or sets a search string that positions the cursor on the first matching row.
+    /// </summary>
+    [Parameter(HelpMessage = "Positions the cursor on the first row matching this regex pattern. Unlike -Filter, non-matching rows remain visible.")]
+    public string? Search { set; get; }
 
     /// <summary>
     ///     Gets or sets the Terminal.Gui driver to use.
     /// </summary>
     [Parameter(HelpMessage =
-        "Sets Terminal.Gui driver to use. Valid values are 'ansi', 'windows', or 'unix'. The default is `ansi`")]
+        "Sets the Terminal.Gui driver to use. Valid values are 'ansi', 'windows', or 'unix'.")]
+    [Alias("ForceDriver")]
     public string? Driver { set; get; }
-
-    /// <summary>
-    ///     Gets or sets a value indicating whether the application should run in full-screen mode
-    ///     using the alternate screen buffer. By default, the application renders inline.
-    /// </summary>
-    [Parameter(HelpMessage =
-        "If specified, the application runs in full-screen mode using the alternate screen buffer. By default, the application renders inline.")]
-    public SwitchParameter FullScreen { set; get; }
 
     /// <summary>
     ///     Gets a value indicating whether the Verbose switch is present.
@@ -94,7 +104,7 @@ public class OutConsoleGridViewCmdletCommand : PSCmdlet, IDisposable
     #endregion Input Parameters
 
     /// <summary>
-    ///     Performs initialization of command execution. Validates that the environment supports grid view.
+    ///     Validates that the environment supports the grid view.
     /// </summary>
     protected override void BeginProcessing()
     {
@@ -125,11 +135,6 @@ public class OutConsoleGridViewCmdletCommand : PSCmdlet, IDisposable
             ProcessObject(InputObject);
     }
 
-    /// <summary>
-    ///     Processes a single object for display in the grid view.
-    /// </summary>
-    /// <param name="input">The PSObject to process.</param>
-    /// <exception cref="FormatException">Thrown when the data type is not supported for Out-ConsoleGridView.</exception>
     private void ProcessObject(PSObject input)
     {
         var baseObject = input.BaseObject;
@@ -149,46 +154,48 @@ public class OutConsoleGridViewCmdletCommand : PSCmdlet, IDisposable
             ThrowTerminatingError(error);
         }
 
-        _psObjects.Add(input);
+        if (!_initialized)
+        {
+            _initialized = true;
+            var applicationData = new ApplicationData
+            {
+                Title = Title,
+                OutputMode = OutputMode,
+                Filter = Filter,
+                MinUI = MinUI,
+                FullScreen = FullScreen,
+                Focus = Focus,
+                Search = Search,
+                Driver = Driver,
+                Verbose = Verbose,
+                Debug = Debug,
+                ModuleVersion = MyInvocation.MyCommand.Version.ToString()
+            };
+            _outConsoleGridView.Initialize(applicationData);
+        }
+
+        _outConsoleGridView.AddObject(input);
     }
 
     /// <summary>
     ///     Performs final processing after all pipeline objects have been received.
-    ///     Displays the console grid view and writes selected objects to the pipeline.
+    ///     Signals the UI that loading is complete and writes selected objects to the pipeline.
     /// </summary>
     protected override void EndProcessing()
     {
         base.EndProcessing();
 
-        // Return if no objects
-        if (_psObjects.Count == 0) return;
+        if (!_initialized) return;
 
-        var applicationData = new ApplicationData
-        {
-            PSObjects = _psObjects.Cast<object>().ToList(),
-            Title = Title ?? "Out-ConsoleGridView",
-            OutputMode = OutputMode,
-            Filter = Filter,
-            MinUI = MinUI,
-            Driver = Driver,
-            FullScreen = FullScreen,
-            Verbose = Verbose,
-            Debug = Debug,
-            ModuleVersion = MyInvocation.MyCommand.Version.ToString()
-        };
-
-        HashSet<int> selectedIndexes = _outConsoleGridView.Run(applicationData);
-
+        HashSet<int> selectedIndexes = _outConsoleGridView.Complete();
         foreach (var idx in selectedIndexes)
         {
-            var selectedObject = _psObjects[idx];
-
-            WriteObject(selectedObject, false);
+            WriteObject(_outConsoleGridView.GetObject(idx), false);
         }
     }
 
     /// <summary>
-    ///     Releases all resources used by the <see cref="OutConsoleGridViewCmdletCommand" />.
+    ///     Releases all resources.
     /// </summary>
     public void Dispose()
     {
