+++ src/Microsoft.PowerShell.ConsoleGuiTools/Properties/launchSettings.json
@@ -30,16 +30,16 @@
       "commandLineArgs": "-NoProfile -NoLogo -NoExit -Command \"& { Import-Module '$(TargetPath)'; Get-Process | Show-ObjectTree -MinUI -filter con }\"",
       "workingDirectory": "$(TargetDir)"
     },
-    "OCTV": {
+    "OCGV": {
       "commandName": "Executable",
       "executablePath": "C:\\Program Files\\PowerShell\\7-preview\\pwsh.exe",
-      "commandLineArgs": "-NoProfile -NoLogo -NoExit -Command \"& { Import-Module '$(TargetPath)'; Get-Process | Select-Object ProcessName, Id,\r\n   @{Name='CPU(s)'; Expression={[math]::Round($_.CPU,2)}},\r\n   @{Name='WorkingSet(MB)'; Expression={[math]::Round($_.WorkingSet64/1MB,2)}},\r\n   @{Name='Private(MB)'; Expression={[math]::Round($_.PrivateMemorySize64/1MB,2)}}\r\n    | Out-ConsoleTableView -Debug -Title \"processes\" }\"",
+      "commandLineArgs": "-NoProfile -NoLogo -NoExit -Command \"& { Import-Module '$(TargetPath)'; Get-Process | Select-Object ProcessName, Id,\r\n   @{Name='CPU(s)'; Expression={[math]::Round($_.CPU,2)}},\r\n   @{Name='WorkingSet(MB)'; Expression={[math]::Round($_.WorkingSet64/1MB,2)}},\r\n   @{Name='Private(MB)'; Expression={[math]::Round($_.PrivateMemorySize64/1MB,2)}}\r\n    | Out-ConsoleGridView -Debug -Title \"processes\" }\"",
       "workingDirectory": "$(TargetDir)"
     },
-    "OCTV Slow": {
+    "OCGV Slow": {
       "commandName": "Executable",
       "executablePath": "C:\\Program Files\\PowerShell\\7-preview\\pwsh.exe",
-      "commandLineArgs": "-NoProfile -NoLogo -NoExit -Command \"& { Import-Module '$(TargetPath)'; 1..50 | ForEach-Object { Start-Sleep -Milliseconds 50; [PSCustomObject]@{ Index = $_; Time = (Get-Date -Format 'HH:mm:ss.fff'); Square = $_ * $_ } }  | Out-ConsoleTableView -Debug }\"",
+      "commandLineArgs": "-NoProfile -NoLogo -NoExit -Command \"& { Import-Module '$(TargetPath)'; 1..50 | ForEach-Object { Start-Sleep -Milliseconds 50; [PSCustomObject]@{ Index = $_; Time = (Get-Date -Format 'HH:mm:ss.fff'); Square = $_ * $_ } }  | Out-ConsoleGridView -Debug }\"",
       "workingDirectory": "$(TargetDir)"
     }
   }
